Update contributors to reflect Twitter changes.
Contribution to the @decredproject Twitter account has evolved over the past few months. A #twitter_ops channel is being used on Slack and Matrix to collaboratively draft and execute project tweets (including retweets). Anyone with an interest in contributing to the Twitter account can ask for an invitation to the channel and can start contributing content and ideas there for evaluation by the Twitter group. As a result, no minority or unilateral veto over tweets is possible. Accordingly, I'm submitting this pull request for consideration and to reflect the change.