Update index.html
Removed "Decentralized Exchange" under "Planned Features".  The other 9 items under this section have either already had a consensus vote or don't require one.  It's my understanding a final decision by the stakeholders whether or not to invest in the development of a DEX is going to be made once PI is released.