Update lottie player.
Noticed lottie animations on `/release` and `/history` pages were not working due to an unhandled exception in lottie player JS.

Tracked it down to the changes in #976 - upgrading Hugo from 0.80 to 0.81 caused the error to occur.

My best guess is this is caused by Hugo updating their version of esbuild. Updating to the latest version of lottie player resolves the problem.