2024 Website Refresh
This is the website refresh described by, and funded by, the proposal
"Decred Website Refresh 2024" (https://proposals.decred.org/record/38a9726).

Only the homepage is changed by this PR, every other page remains
completely unchanged.

All new strings are added in English, i18n translation will files need
to be updated for other languages. The strings in the "Quick Stats"
section are not translatable as the correct layout of this section is
highly dependant on the strings being a predictable/known fixed length.

The numbers in the "Quick Stats" section are populated by minimal
JavaScript which collects the necessary data from https://api.decred.org/.
This section is completely hidden by default, and it is only displayed
if all JavaScript is executed without error.